requirements.txt
fastapi
uvicorn
yt-dlp
opencv-python
imagehash
pillow
numpy
google-genai